Converted to weeks
Converted to weeks, going to split into another branch to remove
wellness and alter the sytem to utilize the same sort of time to death
from AIDS that is used for time to death from non-aids causes.

+++ src/hivMicroSim/Agent/Agent.java
@@ -306,7 +306,7 @@ public abstract class Agent extends OvalPortrayal2D implements Steppable{
      * Will return the base libido prior to any changes caused by disease or other influences on their personality.
      * @return base libido
      */
-    public int getBaseLibido(){
+    public double getBaseLibido(){
         return pp.baseLibido;
     }
     /**
